Reused existing function for fetching diff from URL
