+++ bugbug/phabricator.py
@@ -294,15 +294,16 @@ def get_pending_review_time(rev: RevisionDict) -> timedelta | None:
         return datetime.utcnow() - creation_date
 
 
-def fetch_diff(revision_id, patch_id):
-    try:
-        url = f"https://phabricator.services.mozilla.com/D{revision_id}?id={patch_id}&download=true"
-        response = requests.get(url)
-        response.raise_for_status()
+def fetch_diff_from_url(
+    revision_id, first_patch, second_patch=None, single_patch=False
+):
+    if single_patch:
+        url = f"https://phabricator.services.mozilla.com/D{revision_id}?id={first_patch}&download=true"
+    else:
+        url = f"https://phabricator.services.mozilla.com/D{revision_id}?vs={first_patch}&id={second_patch}&download=true"
+
+    response = requests.get(url)
+    if response.status_code == 200:
         return response.text
-    except requests.HTTPError as e:
-        logger.error(f"HTTP error fetching diff: {e}")
-        return None
-    except Exception as e:
-        logger.error(f"Unexpected error: {e}")
-        return None
+    else:
+        raise Exception(f"Failed to download diff from URL: {url}")
